refactor: move ValidatorMonitor to the BeaconChain
