+++ packages/beacon-node/src/api/impl/validator/index.ts
@@ -1327,7 +1327,7 @@ export function getValidatorApi(
               committeeIndices
             );
             const sentPeers = await network.publishBeaconAggregateAndProof(signedAggregateAndProof);
-            metrics?.onPoolSubmitAggregatedAttestation(seenTimestampSec, indexedAttestation, sentPeers);
+            chain.validatorMonitor?.onPoolSubmitAggregatedAttestation(seenTimestampSec, indexedAttestation, sentPeers);
           } catch (e) {
             const logCtx = {
               slot: signedAggregateAndProof.message.aggregate.data.slot,
@@ -1433,7 +1433,7 @@ export function getValidatorApi(
 
       if (metrics) {
         for (const subscription of subscriptions) {
-          metrics.registerLocalValidator(subscription.validatorIndex);
+          chain.validatorMonitor?.registerLocalValidator(subscription.validatorIndex);
         }
       }
     },
@@ -1470,7 +1470,10 @@ export function getValidatorApi(
 
       if (metrics) {
         for (const subscription of subscriptions) {
-          metrics.registerLocalValidatorInSyncCommittee(subscription.validatorIndex, subscription.untilEpoch);
+          chain.validatorMonitor?.registerLocalValidatorInSyncCommittee(
+            subscription.validatorIndex,
+            subscription.untilEpoch
+          );
         }
       }
     },
