+++ packages/beacon-node/src/node/nodejs.ts
@@ -177,13 +177,7 @@ export class BeaconNode {
       // monitoring relies on metrics data
       opts.monitoring.endpoint
     ) {
-      metrics = createMetrics(
-        opts.metrics,
-        config,
-        anchorState,
-        logger.child({module: LoggerModule.vmon}),
-        metricsRegistries
-      );
+      metrics = createMetrics(opts.metrics, anchorState, metricsRegistries);
       initBeaconMetrics(metrics, anchorState);
       // Since the db is instantiated before this, metrics must be injected manually afterwards
       db.setMetrics(metrics.db);
