+++ packages/beacon-node/test/unit/metrics/utils.ts
@@ -1,12 +1,9 @@
-import {config} from "@lodestar/config/default";
 import {ssz} from "@lodestar/types";
 import {Metrics, createMetrics} from "../../../src/metrics/index.js";
-import {testLogger} from "../../utils/logger.js";
 
 export function createMetricsTest(): Metrics {
   const state = ssz.phase0.BeaconState.defaultViewDU();
-  const logger = testLogger();
-  const metrics = createMetrics({enabled: true, port: 0}, config, state, logger);
+  const metrics = createMetrics({enabled: true, port: 0}, state);
   // we don't need gc metrics running for tests
   metrics.close();
   return metrics;
