+++ packages/state-transition/src/stateTransition.ts
@@ -1,5 +1,5 @@
 import {SLOTS_PER_EPOCH} from "@lodestar/params";
-import {SignedBeaconBlock, SignedBlindedBeaconBlock, Slot, ssz} from "@lodestar/types";
+import {Epoch, SignedBeaconBlock, SignedBlindedBeaconBlock, Slot, ssz} from "@lodestar/types";
 import {toRootHex} from "@lodestar/utils";
 import {BlockExternalData, DataAvailableStatus, ExecutionPayloadStatus} from "./block/externalData.js";
 import {processBlock} from "./block/index.js";
@@ -71,7 +71,8 @@ export function stateTransition(
     executionPayloadStatus: ExecutionPayloadStatus.valid,
     dataAvailableStatus: DataAvailableStatus.available,
   },
-  metrics?: BeaconStateTransitionMetrics | null
+  metrics?: BeaconStateTransitionMetrics | null,
+  validatorMonitor?: IValidatorMonitor | null
 ): CachedBeaconStateAllForks {
   const {verifyStateRoot = true, verifyProposer = true} = options;
 
@@ -90,7 +91,7 @@ export function stateTransition(
 
   // Process slots (including those with no blocks) since block.
   // Includes state upgrades
-  postState = processSlotsWithTransientCache(postState, blockSlot, options, metrics);
+  postState = processSlotsWithTransientCache(postState, blockSlot, options, metrics, validatorMonitor);
 
   // Verify proposer signature only
   if (verifyProposer && !verifyProposerSignature(postState, signedBlock)) {
@@ -145,7 +146,8 @@ export function processSlots(
   state: CachedBeaconStateAllForks,
   slot: Slot,
   epochTransitionCacheOpts?: EpochTransitionCacheOpts & {dontTransferCache?: boolean},
-  metrics?: BeaconStateTransitionMetrics | null
+  metrics?: BeaconStateTransitionMetrics | null,
+  validatorMonitor?: IValidatorMonitor | null
 ): CachedBeaconStateAllForks {
   // .clone() before mutating state in state transition
   let postState = state.clone(epochTransitionCacheOpts?.dontTransferCache);
@@ -157,7 +159,7 @@ export function processSlots(
   // State is already a ViewDU, which won't commit changes. Equivalent to .setStateCachesAsTransient()
   // postState.setStateCachesAsTransient();
 
-  postState = processSlotsWithTransientCache(postState, slot, epochTransitionCacheOpts, metrics);
+  postState = processSlotsWithTransientCache(postState, slot, epochTransitionCacheOpts, metrics, validatorMonitor);
 
   // Apply changes to state, must do before hashing
   postState.commit();
@@ -187,11 +189,24 @@ export function processSlots(
  *   ╚═══════════════════════════════════════════════════════════════════════════════════╝
  *   beacon-node ShufflingCache
  */
+
+type IValidatorMonitor = {
+  registerValidatorStatuses(
+    currentEpoch: Epoch,
+    inclusionDelays: number[],
+    flags: number[],
+    isActiveCurrEpoch: boolean[],
+    isActivePrevEpoch: boolean[],
+    balances?: number[]
+  ): void;
+};
+
 function processSlotsWithTransientCache(
   postState: CachedBeaconStateAllForks,
   slot: Slot,
   epochTransitionCacheOpts?: EpochTransitionCacheOpts,
-  metrics?: BeaconStateTransitionMetrics | null
+  metrics?: BeaconStateTransitionMetrics | null,
+  validatorMonitor?: IValidatorMonitor | null
 ): CachedBeaconStateAllForks {
   const {config} = postState;
   if (postState.slot > slot) {
@@ -219,7 +234,7 @@ function processSlotsWithTransientCache(
 
       const {currentEpoch, inclusionDelays, flags, isActiveCurrEpoch, isActivePrevEpoch, balances} =
         epochTransitionCache;
-      metrics?.registerValidatorStatuses(
+      validatorMonitor?.registerValidatorStatuses(
         currentEpoch,
         inclusionDelays,
         flags,
